requirements.txt
httpx
fastmcp
sse-starlette
starlette
uvicorn
